Add files via upload
The Python project is based on analyzing Unicorn company data. These companies are billion-dollar valued organizations hence the name Unicorn given to them. 
The first step was to import the necessary libraries np, pd, sns, and plt. 
The next step was to read and load the dataset I called thatUnicorn_data. 
The data was then thorougly cleaned and manipulated. 
Univariate analysis was carried out. 
Bivariate and Multivariate analysis was performed as well
Finally, the four recommended analyses were carried and a recommendation was given to improve revenue. 

Challenges Faced
The first challenge faced was finding out the biggest ROI of the unicorn companies. I encountered Value errors and Type errors. As well as converting the billion-dollar values to strings so they can be read in Python. 
The second challenge faced was to find how long it took to become a unicorn I encountered value and type errors here as well.